refactor(scope): eliminate org layer and consolidate into scope (#3872)
Move all org-service functions to scope-member-service with proper
naming (getOrgScope→getScopeWithClerkOrg, getOrganizationStatus→
getScopeMembers, leaveOrganization→leaveScope). Move vm0-admin slug
bypass from /api/org to /api/scope. Delete org-service.ts, /api/org,
and contracts/org.ts.
