+++ turbo/apps/web/app/api/scope/invite/route.ts
@@ -2,7 +2,7 @@ import { NextResponse } from "next/server";
 import { initServices } from "../../../../src/lib/init-services";
 import { getUserId } from "../../../../src/lib/auth/get-user-id";
 import { requireScopeFromRequest } from "../../../../src/lib/scope/resolve-scope";
-import { inviteMember } from "../../../../src/lib/org/org-service";
+import { inviteMember } from "../../../../src/lib/scope/scope-member-service";
 import {
   isBadRequest,
   isNotFound,
